django-backend/requirements.txt
astroid==2.2.5
autopep8==1.4.3
boto3==1.9.191
botocore==1.12.191
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.7
django-allauth==0.39.1
django-cors-headers==2.5.2
django-extensions==2.2.1
django-filter==2.1.0
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-rest-auth==0.9.5
django-storages==1.7.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docutils==0.15
gunicorn==19.9.0
idna==2.8
isort==4.3.16
jmespath==0.9.4
jsonfield==2.0.2
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
oauthlib==3.0.2
psycopg2==2.8.2
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
raven==6.10.0
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
sentry-sdk==0.10.2
six==1.12.0
typed-ast==1.3.1
urllib3==1.25.3
wrapt==1.11.1
